Refactor the admin parser
